requirements.txt
openai
faiss-cpu
numpy
flask
requests
python-dotenv
chromadb
anthropic
langchain
langchain-openai
langchain-anthropic
beautifulsoup4
pypdf
langgraph